+++ pages/index.tsx
@@ -5,7 +5,7 @@ import {
   Bloc1,
   Bloc2,
   Footer,
-  Contact
+  ContactForm
 } from "../components";
 import { fetchProjects } from "../components/Page";
 import { Project } from "../types";
@@ -22,7 +22,7 @@ function Home(props: HomeProps) {
       <Bloc1 />
       <Bloc2 />
       <ProjectsComponent projects={projects} />
-      <Contact />
+      <ContactForm />
       <Footer />
     </Page>
   );
