+++ types.ts
@@ -5,3 +5,9 @@ export type Project = {
   link: string;
   order: number;
 };
+
+export type FormType = {
+  emailInput: string;
+  subjectInput: string;
+  informationInput: string;
+}
